refactor: better separation of concerns
`StaticFileServer` is a module, it describe two state of a server:
`ReadyToRun`, a configured server, with a single `start` method.
And `Running` a running server extending the `Writable` stream of
nodejs, permitting to expose new file by writing into the stream.
When writting to the stream/server, the hash & link is done, and a
notification is made.

`Watcher` has been simplified, it is only responsible to watch new file
in a directory. It extends `Readable` stream.

Like this whenever the readyToRun server is running, we create a watcher
and pipe it to the running server, and voila!