allows configuration files
In order to avoid to many options, a config option is added, this option
permits to specify a file containing the available options in a
`key=value` properties format.

Unfortunately by specifying the `fileToWatch` argument as optional, the
`action` does not match when it is not specified, so its way ugly, but
just use the statefull `program` object.