adds a debouncer to listen to all watch events
- Like this we are able to listen for a file being downloaded/copied
into the watched directory. We will expose it and notify about it only
when no new events are sent during a specified amount of time (default
is 30s, and is configurable using the `--debounce` option).

On tech side:
- A debouncer class has been implemented, extended the `Transform`
stream. The stream is piped between the watcher and the static file server.
- The watcher has been modified to not only send file creation event,
but modification too.
- jest has been added in order to unit test the debouncer.